updated pbtools py
