--- go.mod
@@ -1,14 +1,0 @@
-module github.com/Dav-11/pbtools_lkm
-
-go 1.21.6
-
-require (
-	github.com/spf13/cobra v1.8.0
-	github.com/yoheimuta/go-protoparser/v4 v4.9.0
-	google.golang.org/protobuf v1.33.0
-)
-
-require (
-	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-)
