--- internal/compile/backend/templates/header.tpl
@@ -1,79 +1,0 @@
-/* SPDX-License-Identifier: MIT
- *
- * The MIT License (MIT)
- *
- * Copyright (c) 2019 Erik Moqvist
- *
- * Permission is hereby granted, free of charge, to any person
- * obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without
- * restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
- * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-/**
- * This file was generated by pbtools.
- */
-
-
-#ifndef {{.IncludeGuard}}
-#define {{.IncludeGuard}}
-
-
-#ifndef CHAR_BIT
-#define CHAR_BIT  __CHAR_BIT__
-#endif
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#include "../pbtools.h"
-
-{{range .Includes}}
-#include {{.}}
-{{end}}
-
-/*
- * TYPES
- */
-
-{{range .Types}}
-{{.}}
-{{end}}
-
-/*
- * DECLARATIONS
- */
-
-{{range .Declarations}}
-{{.}}
-{{end}}
-
-/*
- * Internal functions. Do not use!
- */
-
-{{range .InternalDeclarations}}
-{{.}}
-{{end}}
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
