--- internal/compile/model/enum.go
@@ -1,11 +1,0 @@
-package model
-
-type Enum struct {
-	Name          string
-	NameSnakeCase string
-}
-
-type EnumMember struct {
-	NameSnakeCase string
-	FieldNumber   int
-}
