--- internal/compile/model/oneof.go
@@ -1,6 +1,0 @@
-package model
-
-type OneOf struct {
-	FullNameSnakeCase string
-	NameSnakeCase     string
-}
