--- internal/initialize/module_template/.gitignore
@@ -1,15 +1,0 @@
-*.ko
-*.o
-*.o.d
-*.o.cmd
-*.mod.c
-*.mod
-*.mod.cmd
-*.ko.cmd
-**/*.cmd
-modules.order
-Module.symvers
-cmake-build-debug
-.idea
-
-bkp/
