--- internal/initialize/module_template/README.md
@@ -1,22 +1,0 @@
-# Module
-This folder contains the module itself
-
-## Requirements
-- python3
-
-## Install dependencies
-```bash
-make prerequisites
-```
-
-## Generate structs from proto
-```shell
-make generate
-```
-
-## Build
-```shell
-make build
-```
-
-## Lint & static_analysis
