--- internal/initialize/module_template/pbtools/__main__.py
@@ -1,5 +1,0 @@
-# Execute as "python -m cantools"
-
-from . import _main as main
-
-main()
