--- internal/initialize/module_template/pbtools/errors.py
@@ -1,2 +1,0 @@
-class Error(Exception):
-    pass
