--- internal/initialize/module_template/pbtools/version.py
@@ -1,1 +1,0 @@
-__version__ = '0.47.0'
