--- internal/initialize/module_template/proto/hello_world.proto
@@ -1,7 +1,0 @@
-syntax = "proto3";
-
-package hello_world;
-
-message Foo {
-    int32 bar = 1;
-}
