--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/Dav-11/pbtools_lkm/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
