+++ pbtools/__init__.py
@@ -24,9 +24,9 @@ def _main():
 
     # Import when used for less dependencies. For example, curses is
     # not part of all Python builds.
-    from .subparsers import generate_c_source
+    from .subparsers import generate_lkm_source
 
-    generate_c_source.add_subparser(subparsers)
+    generate_lkm_source.add_subparser(subparsers)
 
     args = parser.parse_args()
