+++ test_dir/.gitignore
@@ -1,0 +1,28 @@
+
+# MacOS
+.DS_Store
+
+# IDEs
+**/.idea
+**/.fleet
+**/.vscode
+
+# exe
+**/main
+
+# LKM artifacts
+**/*.ko
+**/*.o
+**/*.o.d
+**/*.o.cmd
+**/*.mod.c
+**/*.mod
+**/*.mod.cmd
+**/*.ko.cmd
+**/*.cmd
+**/modules.order
+**/Module.symvers
+**/cmake-build-debug
+
+# keep go mod
+!**/go.mod
